NewProject/msdocs-python-flask-webapp-quickstart/requirements.txt
Flask==2.0.2
gunicorn
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.0
pyodbc==5.0.1
pypyodbc==1.3.6
azure-storage-blob==12.9.0